feat(MiddlewareFactory): add adjustable log level to logging middleware
This commit enhances the `logging` method in the `MiddlewareFactory` class to accept an optional `logLevel` parameter. This allows the log level for incoming requests to be adjusted as needed. By default, the log level is set to 'info'. The commit also includes unit tests to verify the functionality of the updated `logging` method.

Signed-off-by: Brian Phillips <28457+brianphillips@users.noreply.github.com>